requirements.txt
Django>=3.2,<3.3
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.9.2,<3.0.0
pytz==2021.3
flake8>=4.0.1,<4.1.1
celery>=5.1.2,<5.2.0
django-celery-beat>=2.2.1,<2.3.0
django-celery-results>=2.2.0,<2.3.0
redis==4.0.2
gunicorn==20.1.0